+++ src/transforms/basic.js
@@ -1,7 +1,6 @@
 import {randomLcg} from "d3";
 import {ascendingDefined} from "../defined.js";
 import {arrayify, isOptions, valueof} from "../options.js";
-import {initializer} from "./initializer.js";
 
 // If both t1 and t2 are defined, returns a composite transform that first
 // applies t1 and then applies t2.
@@ -26,12 +25,44 @@ export function basic({
   };
 }
 
+// If both i1 and i2 are defined, returns a composite initializer that first
+// applies i1 and then applies i2.
+export function initializer({
+  filter: f1,
+  sort: s1,
+  reverse: r1,
+  initializer: i1,
+  ...options
+} = {}, i2) {
+  if (i1 === undefined) { // explicit initializer overrides filter, sort, and reverse
+    if (f1 != null) i1 = filterTransform(f1);
+    if (s1 != null && !isOptions(s1)) i1 = composeInitializer(i1, sortTransform(s1));
+    if (r1) i1 = composeInitializer(i1, reverseTransform);
+  }
+  return {
+    ...options,
+    initializer: composeInitializer(i1, i2)
+  };
+}
+
 function composeTransform(t1, t2) {
   if (t1 == null) return t2 === null ? undefined : t2;
   if (t2 == null) return t1 === null ? undefined : t1;
   return function(data, facets) {
     ({data, facets} = t1.call(this, data, facets));
     return t2.call(this, arrayify(data), facets);
+  };
+}
+
+// TODO Don’t export.
+export function composeInitializer(i1, i2) {
+  if (i1 == null) return i2 === null ? undefined : i2;
+  if (i2 == null) return i1 === null ? undefined : i1;
+  return function(data, facets, channels, scales, dimensions) {
+    let c1, d1, f1, c2, d2, f2;
+    ({data: d1 = data, facets: f1 = facets, channels: c1} = i1.call(this, data, facets, channels, scales, dimensions));
+    ({data: d2 = d1, facets: f2 = f1, channels: c2} = i2.call(this, d1, f1, {...channels, ...c1}, scales, dimensions));
+    return {data: d2, facets: f2, channels: {...c1, ...c2}};
   };
 }
 
@@ -43,7 +74,7 @@ export function filter(value, options) {
   return apply(options, filterTransform(value));
 }
 
-export function filterTransform(value) {
+function filterTransform(value) {
   return (data, facets) => {
     const V = valueof(data, value);
     return {data, facets: facets.map(I => I.filter(i => V[i]))};
@@ -54,7 +85,7 @@ export function reverse(options) {
   return {...apply(options, reverseTransform), sort: null};
 }
 
-export function reverseTransform(data, facets) {
+function reverseTransform(data, facets) {
   return {data, facets: facets.map(I => I.slice().reverse())};
 }
 
@@ -66,7 +97,7 @@ export function sort(value, options) {
   return {...apply(options, sortTransform(value)), sort: null};
 }
 
-export function sortTransform(value) {
+function sortTransform(value) {
   return (typeof value === "function" && value.length !== 1 ? sortCompare : sortValue)(value);
 }
